requirements.txt
# ⚠️ for adding new packages, add a locked version here and then run `uv pip install -r requirements.txt`

# Core functionality
python-dotenv==1.0.1  # Environment variables
pynput==1.7.6  # Keyboard shortcuts
sounddevice==0.5.1  # Audio recording
soundfile==0.12.1  # Audio file handling

# AI
litellm==1.63.11 # model routing
tenacity==8.5.0 # Retrying library
openai==1.68.0
anthropic==0.49.0
requests==2.31.0  # For update check

# UI and system interaction
pyautogui==0.9.54  # Cursor/text manipulation
pystray==0.19.5  # System tray icon
Pillow==10.1.0  # Required by pystray for icons
numpy==2.0.2
pyperclip==1.9.0